fix: resolve all build errors and warnings in Server and Tests projects
Agent-Logs-Url: https://github.com/ivanopaulon/EventForge/sessions/b82031d8-a15c-44fb-bed0-9a148896ae11

+++ EventForge.Server/Pages/Dashboard/Logs.cshtml.cs
@@ -50,7 +50,7 @@ public class LogsModel(ILogManagementService logManagementService, ILogger<LogsM
             var result = await logManagementService.GetApplicationLogsAsync(queryParams, HttpContext.RequestAborted);
 
             Logs = result.Items?.ToList() ?? [];
-            TotalCount = result.TotalCount;
+            TotalCount = (int)result.TotalCount;
             TotalPages = (int)Math.Ceiling(TotalCount / (double)PageSize);
             CurrentPage = Math.Max(1, Math.Min(CurrentPage, Math.Max(1, TotalPages)));
         }
